IDAM-146 - removing the default value from the allowlist appsettings because it overrides an empty array passed into it.

+++ NICE.Identity.Authorisation.WebAPI/Repositories/Context.cs
@@ -132,16 +132,31 @@ namespace NICE.Identity.Authorisation.WebAPI.Repositories
         public async Task<int> DeleteUsers(IList<User> users)
         {
             //before removing a user, we also need to remove the UserAcceptedTermsVersion, Job and UserRole for the user, if they exist.
+            if (!users.Any())
+            {
+	            return 0;
+            }
 
             var userIds = users.Select(user => user.UserId);
 
             var userRolesForUsers = UserRoles.Where(ur => userIds.Contains(ur.UserId));
-            var jobsForUsers = Jobs.Where(job => userIds.Contains(job.UserId));
-            var acceptedTermsForUsers = UserAcceptedTermsVersions.Where(uatv => userIds.Contains(uatv.UserId));
+            if (userRolesForUsers.Any())
+            {
+	            UserRoles.RemoveRange(userRolesForUsers);
+            }
 
-            UserRoles.RemoveRange(userRolesForUsers);
-            Jobs.RemoveRange(jobsForUsers);
-            UserAcceptedTermsVersions.RemoveRange(acceptedTermsForUsers);
+            var jobsForUsers = Jobs.Where(job => userIds.Contains(job.UserId));
+            if (jobsForUsers.Any())
+            {
+	            Jobs.RemoveRange(jobsForUsers);
+            }
+
+            var acceptedTermsForUsers = UserAcceptedTermsVersions.Where(uatv => userIds.Contains(uatv.UserId));
+            if (acceptedTermsForUsers.Any())
+            {
+	            UserAcceptedTermsVersions.RemoveRange(acceptedTermsForUsers);
+            }
+
             Users.RemoveRange(users);
             
             return await SaveChangesAsync();
